Restructured to use an abstract instead of an interface

+++ GradeBook/Student.cs
@@ -7,7 +7,7 @@ namespace GradeBook
         string Name { get; set; }
         StudentType Type { get; set; }
         EnrollmentType Enrollment { get; set; }
-        List<int> Grades { get; set; }
+        List<decimal> Grades { get; set; }
 
         public Student()
         {
